requirements.txt
Flask==2.2.0
Flask-SQLAlchemy==2.5.1
Flask-RESTful==0.3.9
SQLAlchemy==1.4.15
flask-marshmallow==1.1.0
marshmallow==3.12.1
marshmallow-sqlalchemy==0.25.0
Werkzeug==2.2.2
Jinja2==3.1.2
itsdangerous==2.1.2
psycopg2-binary==2.9.5
PyJWT==2.1.0
Flask-JWT==0.2.0
Flask-JWT-Extended==4.6.0
redis==3.5.3
celery==5.2.3
moviepy==1.0.3
pillow==10.2.0
tzdata
psycopg2==2.9.3
google-cloud-storage==2.16.0
google-cloud-pubsub==2.19.0